+++ Changelog.markdown
@@ -10,6 +10,7 @@
 * __upd__:  updated regular expressions used for option string recognition,
 
 * __new__:  `clap/base.py` module,
+* __new__:  `clap/checker.py` module,
 * __new__:  first tests added for versions `0.6.x`,
 
 * __rem__:  `hint` parameter is removed from all CLAP components,